sp モードの表示分け
